Use scoord coordinates to get area and length
